Added type hints, cleaned up imports, and cleaned up redundant file Slots.py
