+++ MediaControls.py
@@ -1,0 +1,13 @@
+class Media_Controls():
+    def toggle_play_pause(self):
+        print("toggled")
+
+    def skip(self):
+        #interact with song queue
+        print("song skipped")
+
+    def previous(self):
+        #if current timestamp <5s
+        print("previous song loaded")
+        #else:
+            #print("song restarted") // set time to 0
