+++ Songs.py
@@ -1,7 +1,7 @@
 from PySide6.QtWidgets import QLabel, QListWidget, QListWidgetItem
 
 class Song(QListWidgetItem):
-    def __init__(self, path_to_file, title, artist): #TODO add duration, tags[], playlistsIn[].. ,maybe other key fields
+    def __init__(self, path_to_file, title, artist): #TODO add duration, tags[], collectionsIn[].. ,maybe other key fields
         super().__init__()
         self.path_to_file = path_to_file
         self.title = title
@@ -15,7 +15,7 @@ class Songs(QListWidget):
         self.label = QLabel("Songs")
         self.setVisible(True)
 
-    def populate(self, songList):
+    def populate(self, songList: list['Song']):
         for row, item in enumerate(songList):
             item.setText(item.title)
             self.insertItem(row, item)
