+++ main.py
@@ -1,6 +1,5 @@
-from PySide6.QtGui import QIcon
-import slots
-from slots import *
+import sys
+from PySide6.QtWidgets import QApplication, QMainWindow
 from MainWindow import MainWindow
 
 def main():
@@ -8,14 +7,10 @@ def main():
 #
 # App
     app = QApplication(sys.argv)
-    # Run the main Qt loop
-    #engine = QQmlApplicationEngine()
-    #engine.load( QUrl("qml/UI_desktop.qml"))
 
+    # Run the main Qt loop
     window = MainWindow()
     window.show()
-
-
     sys.exit(app.exec())
 
 if __name__ == '__main__':
